feat: Use loopback (#26)
When using ports instead of sockets, the Neovim port (both remote and
client) are bound to the loopback address, rather than `0.0.0.0`. This
should be a little more secure.

Ports tunneled via `NvrhTunnelPort` are still bound to `0.0.0.0` on the
client machine, that's because I still need to use my phone / tablet to
check things out sometimes, and if they were bound to `127.0.0.1` I
wouldn't be able to connect. Ideally, this would be configurable at
mapping-time, something like `NvrhTunnelPort 3000 interface=all` /
`NvrhTunnelPort 3000 interface=loopback`

- Clean up local socket
- Clean up remote socket
- Clean up remote browser script
- Use `filepath.Join` instead of `path.Join`
- Extract logger setup