feat: Better multiple sessions (#41)
Closes #40 

Note: Since each nvim client doesn't know which nvrh client actually
initiated a command, things like ports and browsers are opened on all
machines connected to the session.

## Preview

In this video I'm starting a session in macOS, then reconnecting to that
session in Linux.

Once reconnecting, any ports that were previously mapped are mapped
there as well, so the browser can just be opened directly.



https://github.com/user-attachments/assets/6b21b2c0-41c4-4b44-b41c-b893c3f8fa02