feat: Extract lua code (#44)
Been wanting to do this for a while, writing any language in a file that
is not of that language is always a pain point.

So now the lua lives in its own place, and can be type checked. The
functionality is still accessible via `_G._nvrh`, I had started making
most of the API private when doing this, but undid those changes just to
get this in and working sooner.