feat: Windows terminal support (#55)
So they key here is `start "" /WAIT nvim ...`, but getting there is
annoying.

We write a `.bat` file, because the user's shell might be git-bash which
would need translating windows paths and cmd-style arguments. We still
have to be aware of powershell and cmd, so that's all plumbed through,
so powershell users on Linux should also be supported now.

Closes #54 

## Preview



https://github.com/user-attachments/assets/908ddee1-ff12-4a89-ba7c-fc35039cb38f